feat: Implement error handling and validation in services, add ErrorResponse class, and enhance UserRepo with username lookup

+++ social-media-be/src/main/java/org/example/socialmediabe/service/CommentService.java
@@ -1,9 +1,12 @@
 package org.example.socialmediabe.service;
 
 import jakarta.transaction.Transactional;
+import jakarta.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import org.example.socialmediabe.model.Comment;
 import org.example.socialmediabe.repository.CommentRepo;
+import org.example.socialmediabe.repository.PostRepo;
+import org.example.socialmediabe.repository.UserRepo;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,8 +15,23 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentService {
     private final CommentRepo commentRepo;
+    private final PostRepo postRepo;
+    private final UserRepo userRepo;
 
     public Comment addComment(Comment comment) {
+        //if (post exists, user exists)
+        if(comment.getPost() == null || comment.getPost().getId() == null) {
+            throw new ValidationException("Post cannot be null");
+        }
+        if(postRepo.findById((long) comment.getPost().getId()).isEmpty()) {
+            throw new ValidationException("Post does not exist");
+        }
+        if(comment.getUser() == null || comment.getUser().getId() == null) {
+            throw new ValidationException("User cannot be null");
+        }
+        if(userRepo.findById(comment.getUser().getId()).isEmpty()) {
+            throw new ValidationException("User does not exist");
+        }
         return commentRepo.save(comment);
     }
 
@@ -22,6 +40,10 @@ public class CommentService {
     }
 
     public void deleteComment(Long id) {
+        //if(comment exists)
+        if(commentRepo.findById(id).isEmpty()) {
+            throw new ValidationException("Comment does not exist");
+        }
         commentRepo.deleteById(id);
     }
 }
